Move deposit_through_contract_call back to sdk::deposits
