+++ sdk/src/l2/deposit.rs
@@ -9,7 +9,7 @@ use ethrex_rpc::{
     EthClient,
     clients::{EthClientError, Overrides},
 };
-use ethrex_sdk::calldata::encode_calldata;
+use ethrex_sdk::{calldata::encode_calldata, get_address_from_secret_key};
 use keccak_hash::H256;
 use secp256k1::SecretKey;
 
@@ -23,6 +23,44 @@ pub async fn deposit_through_transfer(
     eth_client: &EthClient,
 ) -> Result<H256, EthClientError> {
     transfer(amount, from, bridge_address, from_pk, eth_client).await
+}
+
+pub async fn deposit_through_contract_call(
+    amount: U256,
+    to: Address,
+    depositor_private_key: &SecretKey,
+    bridge_address: Address,
+    eth_client: &EthClient,
+) -> Result<H256, EthClientError> {
+    let l1_from = get_address_from_secret_key(depositor_private_key)?;
+    let calldata = encode_calldata("deposit(address)", &[Value::Address(to)])?;
+    let gas_price = eth_client
+        .get_gas_price_with_extra(20)
+        .await?
+        .try_into()
+        .map_err(|_| {
+            EthClientError::InternalError("Failed to convert gas_price to a u64".to_owned())
+        })?;
+
+    let deposit_tx = eth_client
+        .build_generic_tx(
+            TxType::EIP1559,
+            bridge_address,
+            l1_from,
+            calldata.into(),
+            Overrides {
+                from: Some(l1_from),
+                value: Some(amount),
+                max_fee_per_gas: Some(gas_price),
+                max_priority_fee_per_gas: Some(gas_price),
+                ..Default::default()
+            },
+        )
+        .await?;
+
+    let signer = Signer::Local(LocalSigner::new(*depositor_private_key));
+
+    send_generic_transaction(eth_client, deposit_tx, &signer).await
 }
 
 pub async fn deposit_erc20(
